pyproject.toml
[project]
name = "office2llm"
version = "0.1.0"
description = "Convert Office docs and PDFs to per-page PNGs (LibreOffice + PDFium) + and per-page OCR text for advanced document understanding."
requires-python = ">=3.10"
dependencies = [
  "pypdfium2==5.5.0",
  "Pillow==12.1.1",
  "google-genai==1.65.0",
]

[project.scripts]
office2llm = "office2llm:cli"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

